fix: cast NULL to int in brand filter option UNION (#183)
* fix: restore @swc/helpers@0.5.21 nested entry for npm 10.x compat

Docker's node:20-alpine ships npm 10.8.2 which requires this nested
entry under next-intl/node_modules/ to satisfy @swc/core's optional
peer dep. Newer npm versions silently resolve it but 10.x fails on
npm ci without it.

* fix: cast NULL to int in brand filter option UNION for type compatibility

+++ src/datapulse/analytics/repository.py
@@ -139,7 +139,7 @@ class AnalyticsRepository:
             ) cats
             UNION ALL
             SELECT * FROM (
-                SELECT 'brand', drug_brand, NULL
+                SELECT 'brand', drug_brand, NULL::int
                 FROM public_marts.agg_sales_by_product
                 WHERE drug_brand IS NOT NULL
                 GROUP BY drug_brand
